sns: hare issues duplicate sns repair trigger
Hare uses motr spiel api to trigger sns recovery operations. Presently,
Hare adds the sns operation to broadcast queue, this was changed as part
of commit 7543caf, but since spiel api connects to all the ioservices
in the cluster and sends the trigger to all the ioservices itself, hare,
further invoking the spiel command on all the nodes leads to duplication
of this trigger.

Solution:
Revert to commit `bc44695` and let only rc invoke the spiel command
for sns operations.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>