broadcast: process state may not be updated on node failure
If a node fails, remote hax process will get a notification from Consul/HA,
hax, if its a RC (recovery coordinator) node will update the state of the
remote processes corresponding to the node in Consul KV. But it is presently
skipping the broadcast of notification to motr processes if the node is not rc.

Solution:
Even if the hax process is not a rc, the failure must still be a broadcast
to the motr processes.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>