ha: process state update is not a broadcast
It is possible that Consul may not send service state notifications
periodically as configured, see hashicorp/consul#7069.
Sometimes the notification is sent only on Consul service state change.
During bootstrap if a peer process has not started, Hare reports it as offline
in nvec reply to another process. Relying on Consul notification, Hare expects
that the offline state will be updated to online on receiving a corresponding
periodical Consul notification. But as the notification might get delayed or
not arrive at all, the offline state of the process may not be updated to
online, which may further lead to inconsistencies.

Solution:
- Watch `processes/` key value for the updates reported by the individual
process and accordingly broadcast the process state change to peer motr
processes.
- Handle delays in Consul notifications.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>