CORTX-29668: state of the drive under recovery changes automatically
It may happen that a process was intermittently reported as offline and
reported as online subsequently. Presently, Hare just skips the drive
in `failed` state and ignores `repairing/repaired/rebalancing` state of
the drive. Thus If a drive is in the other recovery states then it is
possible that hare may ignore that and transition the drive to online
in such a case.

Solution:
Consider recovery states (`repairing/repaired/rebalancing`) as well along
with the `failed` state for drives during state update.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>