CORTX-29700: function to update motr process state is complicated
Presently the handler function to update and broadcast motr
process state changes is a bit convulated and is difficult to
read. Also, hax updates the remote process's state only in
case of node failure but not if process is reported as offline.

Solution:
- Simplify process state update logic.
- Update remote process state if the current state of the
process is either failed or offline.
- Add M0_CONF_HA_PROCESS_HA to process types
- Handle unavailability of consul kvs during bootstrap,
retry if an expected consul kv is unavailable.
- Handle unavailbility of motr fids during hax startup by
retrying in case of failure.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>