CORTX-29944: Bytecount thread gets aborted when receiving error from motr API
During node failures, failed services are falsely reported as active
due to delay issue by Consul.
When we try to connect those service we get an error from motr apis.
Instead of aborting the thread, we should keep the thread running but
break the iteration and start a new iteration.
Also log the error reported.

Signed-off-by: Shreya Karmakar <shreya.karmakar@seagate.com>