CORTX-29199: Add support for DIX repair and rebalance in service types
It is observed that, on cluster, currently no DIX repair and rebalance services are enabled even after adding DIX pool in the solutions.yaml or CDF file.
Same is observed using consul kv get command.
Because of this, whenever DIX repair or rebalance for is triggered, the motr-server is not able to lookup the respective opcode as those are not registered. Hence it returns protocol error (-EPROTO).

Solution:
The service types for DIX repair and rebalance need to be enabled in cfgen file.

Signed-off-by: ‘Swapnil <swapnil.chaudhary@seagate.com>