ha: consul agent restart may lead to inconsistencies
In container environment, Hare starts hax and a consul client
agent in the Hare container. On Pod restart, typically, Pods
ip address changes. This leads to inconsistencies as Consul
sees different ip addreses for the same node name. Furthermore,
the restarted Consul agent fails to join the cluster and rest
of the cluster sees it a failed. This causes Hare to report
the corresponding node and motr services as offline.

Solution:
- Generate Consul node name dynamically during every consul
restart. This mean it will be altogether a new consul agent
with new ip address. This allows the restarted agent to
join the cluster as new agent and in-turn helps the correspon-
ding node to be reported as online by Hare.
- Handle related changes with respect to Consul node name,
  - elect-rc-leader must use the consul node name for session
    creation.
  - Fix hctl bootstrap.
  - Save node-name to consul node-name mapping in Consul.
- Use hax node hostname instead of ipaddress to start hax
http server.
- Fix server unit tests to use hostname instead of ipaddresses.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>