CORTX-29199: DIX repair and rebalance service types are missing
It is observed that, on cluster, currently no DIX repair and rebalance services
are enabled even after adding DIX pool in the solutions.yaml or CDF file.
Same is observed using consul kv get command.
Because of this, whenever DIX repair or rebalance operations are triggered,
corresponding service fop types are missing as the respective services are
not registered. Thus, corresponding operation fails with and -EPROTO error.

Solution:
Enable DIX repair and rebalance service types in cfgen in-order to configure
the same during motr configuration.

Signed-off-by: ‘Swapnil <swapnil.chaudhary@seagate.com>