CORTX-29982: Incorrect list-item indent
CORTX-29982: Incorrect list-item indent

Solution:
Add 2 spaces at the start of the statements in documentation.

Signed-off-by: hessio patrick.hession@seagate.com

mini-prov: consul agent may not leave on termination

It is possible that the Hare node might restart and get a new ip address
on restart. It takes some time for the Consul cluster to evict such a
Consul agent with stale ipaddress, this mainly happens if the Consul agent
does not leave the cluster. Furthermore, this may lead to connection
failures during the Consul agent startup.

Solution:
- Use specific bind address to start a consul agent
- Advertise consul agent bind address so that peers would get to know
  about its new address if changed.
- add node-id and node name to consul parameter.
- retry if hax server configuration fails during bootstrap.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>

ha: consul agent restart may lead to inconsistencies

In container environment, Hare starts hax and a consul client
agent in the Hare container. On Pod restart, typically, Pods
ip address changes. This leads to inconsistencies as Consul
sees different ip addreses for the same node name. Furthermore,
the restarted Consul agent fails to join the cluster and rest
of the cluster sees it a failed. This causes Hare to report
the corresponding node and motr services as offline.

Solution:
- Generate Consul node name dynamically during every consul
restart. This mean it will be altogether a new consul agent
with new ip address. This allows the restarted agent to
join the cluster as new agent and in-turn helps the correspon-
ding node to be reported as online by Hare.
- Handle related changes with respect to Consul node name,
  - elect-rc-leader must use the consul node name for session
    creation.
  - Fix hctl bootstrap.
  - Save node-name to consul node-name mapping in Consul.
- Use hax node hostname instead of ipaddress to start hax
http server.
- Fix server unit tests to use hostname instead of ipaddresses.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>

Fixes Codacy issue "Incorrect list-item indent: add 2 spaces"

Signed-off-by: hessio <patrick.hession@seagate.com>

Fixes Codacy issue 'Incorrect list-item indent: add X spaces'

Signed-off-by: hessio <patrick.hession@seagate.com>