Add unit test simulating single node failure
This is a unit test to check behavior on single node failure.
We invoke the handler for a state update indicating a process is
gone, then check that the appropriate components (drive, controller,
node, etc.) are marked failed in Consul. We also check that the
failures are broadcast through hax.

This test only covers a particular code path (total node failure).
There are other code paths for different scenarios (e.g. drive
failure or service crash) that will require additional unit tests.

Signed-off-by: Tim Shaffer <tim.shaffer@seagate.com>