Run m0tracedump when creating support bundles
Motr traces in the support bundles are currently in binary form,
which can only be parsed by the exact Motr version that created
them. A YAML format is also available that's easier to work with.
Motr currently generates YAML traces in its support bundles (see
https://github.com/Seagate/cortx-motr/blob/ae3c398111ffae336515613aacd22e89c914683c/utils/m0reportbug#L754
for details). This adds a similar step to Hare's support bundle
generation process.

Signed-off-by: Tim Shaffer <tim.shaffer@seagate.com>