server: consul sends duplicate events for already processed events
As hare configures watcher for entire `processes` Consul KV tree,
Consul sends notification for entire tree even if subset of keys change.
This adds to redundant processing of already processed events.

Solution:
Maintain events cache in hax server and use the same to verify if a
consul event is duplicate or not. Create and send Broadcasts events
only for the actual events.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>