Problem: hax may not report a motr failure in real time (#1896)
Hax may receive a Motr/S3server failure notification from Consul or
cortx-ha. But the notification can be delayed and the process might
have already restarted. When a Motr/S3server process, it requests
entrypoint information from Hare, which implies that the service is
starting. This can be used as a sure point to notify a failure before
replying to the process's entrypoint request.
Also, to note that it is important to identify the bootstrap situation
against the failure case. We need not report failure for a bootstrapping
process.

Solution:
- Maintain a motr/s3server restart count, fetch the count when hax gets
a First entrypoint request. Check if the process is starting for the
first time or subsequent.
- If a motr/s3server process is not starting for the first time, send
a failure notification for that process before replying to the entrypoint
request.
- Update the process restart count on receiving entrypoint request.
- Update process status in consul kv to stopped to broadcast failure to
  all the nodes on FirstEntrypoint request.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>