Problem: hax entrypoint reply may delay on restart (#2124)
1. Hare sends OFFLINE event on whenever process restarts, it does not
check if the process is a hax or motr client process. If hax process
restarts, it sends FirstEntrypointRequest and will try to send offline
for itself, since no halink is established, the delivery will timeout
and can further delay the entrypoint processing. Which may cause other
side effects. There's no need to send OFFLINE for hax and motr client
processes.
2. Since the halink corresponding to the restarted motr process is not
disconnected yet, haxtries to send the notification to the restarted
process as well. This causes hax timeout waiting for delivery
confirmation and must be avoided.

Solution:
1. Skip Hax and motr client processes while sending OFFLINE for
FirstEntrypoint requests.
2. Add a way to skip some processes from the notification list.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>