CORTX-32741: HARE - Consul Support for GCONF
Problem:
Provisioner team is planning to migrate file based Gconf to consule based Gconf.
In order to do this, hare component needs some changes to support this migration.

Solution:
There are changes required in way we fetch the keys and its values from Gconf
 e.g. endpoints, Hostname, machineIDs.
Approach we took is,
a) for any key we want to fetch, first we need to fetch num_XXX key from Gconf.
b) based on this number, we need to iterate over each key and get desired value of key.

Signed-off-by: pavankrishnat <pavan.k.thunuguntla@seagate.com>