CORTX-33628: Hare ha events are not ordered
Hare monitors motr and rgw processes for failures and
broadcasts the same to the entire cluster. But presently
they are not ordered and there's not way to determine the
order in which they were processed across the cluster.
Out of order events affects IO post or during failures
as events are not delivered properly.
Worst, it is extremely difficult to debug the same through
several logs.
Thus, it is important to maintain the event order and logging.

Solution:
- Use broadcast queue to broadcast events and maintain order.
- Let RC only handle the node failure events.
- Let local node handle the process failure event.
- Increase consul service watcher interval and set timeout
  for probe to avoid frequent Consul watcher events.
- Guard from delayed Consul events by checking process
  node availability and allowing process local node to
  process the event if the corresponding node's alive.
- Notify consistent state for entire process tree.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>