CORTX-29871: process state machine update multiple times
Process events are a broadcast to all the nodes in the cluster.
Not all the nodes are required to update the entire process configuration
tree in KV. Only the process's local hax or RC must do that.

Solution:
- Avoid updating process configuration tree in KV, but notify local motr
processes about any remote process state updates.
- Set HA and Confd process states to M0_CONF_HA_PROCESS_RECOVERED
on receiving M0_CONF_HA_PROCESS_STARTED.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>