CORTX-33789: Hare Config stage fail with Motr Clients
Problem:
If num_client_inst is specified a non-zero value, deployment is failing.
Issue is with search key. For data nodes we used 'io':'motr', it returns the
list of machine IDs which includes both data node and motr client nodes,
because of which deployment was failing.

Solution:
Separated search key for data node and motr client node.

Signed-off-by: pavankrishnat <pavan.k.thunuguntla@seagate.com>