CORTX-33705: confd and ios ports are not aligned with cluster (#2149)
Problem:
The confd and ios Port information in hctl status is not matching with
cluster.conf in both Both basic Linux and K8s

In basic Linux, ios Port in hctl status is 21003 instead of 21001
(as per cluster.conf). And in K8s, both confd and ios Port are not matching
with cluster.conf

Solution:
Modified ios and confd port info to fetch default port.
For K8s, just by changing order in m0serverT list fixes this issue. But to fix
in basic Linux, updated the default ports and avoid rounding off to fetch
actual ports from cluster.conf.

Signed-off-by: pavankrishnat <pavan.k.thunuguntla@seagate.com>

Signed-off-by: pavankrishnat <pavan.k.thunuguntla@seagate.com>