CORTX-28971: update service fids along with process fids (#2153)
A motr client restart is considered as a permanent failure
and thus, if an existing motr client restarts its existing
fid is updated by Hare. A motr process comprises of motr
services associated to it. Interested motr modules subscribe
to change in the service states, also, motr confc (configuration
cache) maintains a hierarchy of processes and services, thus
if a process fid changes without updating the service fids,
it will not be possible to map a given process to its corresponding
services. Thus, It is important to update the services fids
along with its respective process fid.

Solution:
- Allocate service fids along with its corresponding process fid.
- Make fid allocation routine generic to work for any motr
  configuration object.

Signed-off-by: Mandar Sawant <mandar.sawant@seagate.com>